requirements.txt
astroid==2.1.0
autopep8==1.4.3
Django==2.1.7
djangorestframework==3.9.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pep8==1.7.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.2.2
pytz==2018.9
six==1.12.0
typed-ast==1.3.1
wrapt==1.11.1
